requirements.txt
tqdm==4.67.1
numpy==2.1.3
pandas==2.2.3
jiwer==3.1.0
transformers==4.48.2
torch==2.6.0
torchaudio~=2.0.0
speechbrain==1.0.2
librosa==0.10.2.post1
scipy==1.15.1
scikit-learn==1.6.1
praat-parselmouth==0.4.5
soundfile==0.13.1
urllib3==1.26.20
pypinyin==0.44.0
pyloudnorm==0.1.1
numba==0.61.0
cvxopt==1.3.2
sounddevice==0.4.7
matplotlib==3.10.0
phonemizer==3.2.1
dragonmapper==0.3.0
PyYAML==6.0.2
openpyxl==3.1.5
seaborn==0.13.2
noisereduce==3.0.3
accelerate==1.3.0